clean up code and switch to new code
