+++ imodels/importance/__init__.py
@@ -2,7 +2,10 @@
 Feature importance methods for black box models
 """
 
-from .gmdi import GMDI, GmdiHelper
-from .ppms import RidgePPM, LogisticPPM, RobustPPM, GlmPPM, GenericPPM
+from .rf_plus import RandomForestPlusRegressor, RandomForestPlusClassifier
+from .gmdi import ForestGMDI, TreeGMDI
+from .ppms import GenericRegressorPPM, GenericClassifierPPM, \
+    GlmRegressorPPM, GlmClassifierPPM, RidgeRegressorPPM, RidgeClassifierPPM, \
+    LogisticClassifierPPM, RobustRegressorPPM, LassoRegressorPPM
 from .block_transformers import IdentityTransformer, TreeTransformer, \
     CompositeTransformer, GmdiDefaultTransformer
