+++ imodels/importance/block_transformers.py
@@ -4,6 +4,7 @@ import numpy as np
 from collections import defaultdict
 
 from sklearn.ensemble import BaseEnsemble
+from sklearn.ensemble._forest import _generate_unsampled_indices, _generate_sample_indices
 
 from .local_stumps import make_stumps, tree_feature_transform
 
@@ -174,6 +175,25 @@ class BlockPartitionedData:
         return all_data
 
     def train_test_split(self, train_indices, test_indices):
+        """
+        Split the data intro training and test partitions given the
+        training and test indices. Return the training and test
+        block partitioned data objects.
+
+        Parameters
+        ----------
+        train_indices: array-like of shape (n_train_samples,)
+            The indices corresponding to the training samples
+        test_indices: array-like of shape (n_test_samples,)
+            The indices corresponding to the training samples
+
+        Returns
+        -------
+        train_blocked_data: BlockPartitionedData
+            Returns the training block partitioned data set
+        test_blocked_data: BlockPartitionedData
+            Returns the test block partitioned data set
+        """
         train_blocks = [self.get_block(k)[train_indices, :] for
                         k in range(self.n_blocks)]
         train_blocked_data = BlockPartitionedData(train_blocks)
@@ -193,10 +213,64 @@ class BlockTransformerBase(ABC):
     features for each original feature
     """
 
+    def __init__(self):
+        self._centers = {}
+        self._scales = {}
+        self.is_fitted = False
+
+    def fit(self, X):
+        """
+        Fit (or train) the block transformer using the data matrix X.
+
+        Parameters
+        ----------
+        X: ndarray
+            The data matrix to be used in training
+        """
+        for k in range(X.shape[1]):
+            self._fit_one_feature(X, k)
+        self.is_fitted = True
+
+    def check_is_fitted(self):
+        """
+        Check if the transformer has been fitted. Returns an error if not
+        previously fitted.
+        """
+        if not self.is_fitted:
+            raise AttributeError("Transformer has not yet been fitted.")
+
     def transform_one_feature(self, X, k, center=True, normalize=False):
         """
         Obtain a block of engineered features associated with the original
-        feature with index k.
+        feature with index k using the (previously) fitted transformer.
+
+        Parameters
+        ----------
+        X: ndarray
+            The data matrix to be transformed
+        k: int
+            Index of feature in X to be transformed
+        center: bool
+            Flag for whether to center the transformed data
+        normalize: bool
+            Flag for whether to rescale the transformed data to have unit
+            variance
+
+        Returns
+        -------
+        data_block: ndarray
+            The block of engineered features associated with the original
+            feature with index k.
+        """
+        data_block = self._transform_one_feature(X, k)
+        data_block = self._center_and_normalize(data_block, k, center, normalize)
+        return data_block
+
+    def transform(self, X, center=True, normalize=False):
+        """
+        Transform a data matrix into a BlockPartitionedData object comprising
+        one block for each original feature in X using the (previously) fitted
+        trasnformer.
 
         Parameters
         ----------
@@ -210,19 +284,48 @@ class BlockTransformerBase(ABC):
 
         Returns
         -------
+        blocked_data: BlockPartitionedData object
+            The transformed data
+        """
+        self.check_is_fitted()
+        n_features = X.shape[1]
+        data_blocks = [self.transform_one_feature(X, k, center, normalize) for
+                       k in range(n_features)]
+        blocked_data = BlockPartitionedData(data_blocks)
+        return blocked_data
+
+    def fit_transform_one_feature(self, X, k, center=True, normalize=False):
+        """
+        Fit the transformer and obtain a block of engineered features associated with
+        the original feature with index k using this fitted transformer.
+
+        Parameters
+        ----------
+        X: ndarray
+            The data matrix to be fitted and transformed
+        k: int
+            Index of feature in X to be fitted and transformed
+        center: bool
+            Flag for whether to center the transformed data
+        normalize: bool
+            Flag for whether to rescale the transformed data to have unit
+            variance
+
+        Returns
+        -------
         data_block: ndarray
             The block of engineered features associated with the original
             feature with index k.
-
         """
-        data_block = self._transform_one_feature(X, k)
-        data_block = _center_and_normalize(data_block, center, normalize)
+        data_block = self._fit_transform_one_feature(X, k)
+        data_block = self._center_and_normalize(data_block, k, center, normalize)
         return data_block
 
-    def transform(self, X, center=True, normalize=False):
+    def fit_transform(self, X, center=True, normalize=False):
         """
-        Transform a data matrix into a BlockPartitionedData object comprising
-        one block for each original feature in X
+        Fit the transformer and transform a data matrix into a BlockPartitionedData
+        object comprising one block for each original feature in X using this
+        fitted transformer.
 
         Parameters
         ----------
@@ -240,14 +343,34 @@ class BlockTransformerBase(ABC):
             The transformed data
         """
         n_features = X.shape[1]
-        data_blocks = [self.transform_one_feature(X, k, center, normalize) for
+        data_blocks = [self.fit_transform_one_feature(X, k, center, normalize) for
                        k in range(n_features)]
         blocked_data = BlockPartitionedData(data_blocks)
+        self.is_fitted = True
         return blocked_data
+
+    @abstractmethod
+    def _fit_one_feature(self, X, k):
+        pass
 
     @abstractmethod
     def _transform_one_feature(self, X, k):
         pass
+
+    def _fit_transform_one_feature(self, X, k):
+        self._fit_one_feature(X, k)
+        return self._transform_one_feature(X, k)
+
+    def _center_and_normalize(self, data_block, k, center=True, normalize=False):
+        if center:
+            data_block = data_block - self._centers[k]
+        if normalize:
+            if any(self._scales[k] == 0):
+                raise Warning("No recaling done."
+                              "At least one feature is constant.")
+            else:
+                data_block = data_block / self._scales[k]
+        return data_block
 
 
 class IdentityTransformer(BlockTransformerBase, ABC):
@@ -255,6 +378,10 @@ class IdentityTransformer(BlockTransformerBase, ABC):
     Block transformer that creates a block partitioned data object with each
     block k containing only the original feature k.
     """
+
+    def _fit_one_feature(self, X, k):
+        self._centers[k] = np.mean(X[:, [k]])
+        self._scales[k] = np.std(X[:, [k]])
 
     def _transform_one_feature(self, X, k):
         return X[:, [k]]
@@ -279,6 +406,7 @@ class TreeTransformer(BlockTransformerBase, ABC):
     """
 
     def __init__(self, estimator, data=None):
+        super().__init__()
         self.estimator = estimator
         self.oob_seed = self.estimator.random_state
         # Check if single tree or tree ensemble
@@ -302,8 +430,19 @@ class TreeTransformer(BlockTransformerBase, ABC):
             self.stumps[stump.feature].append(stump)
         self.n_splits = {k: len(stumps) for k, stumps in self.stumps.items()}
 
+    def _fit_one_feature(self, X, k):
+        stump_features = tree_feature_transform(self.stumps[k], X)
+        self._centers[k] = np.mean(stump_features, axis=0)
+        self._scales[k] = np.std(stump_features, axis=0)
+
     def _transform_one_feature(self, X, k):
         return tree_feature_transform(self.stumps[k], X)
+
+    def _fit_transform_one_feature(self, X, k):
+        stump_features = tree_feature_transform(self.stumps[k], X)
+        self._centers[k] = np.mean(stump_features, axis=0)
+        self._scales[k] = np.std(stump_features, axis=0)
+        return stump_features
 
 
 class CompositeTransformer(BlockTransformerBase, ABC):
@@ -325,8 +464,8 @@ class CompositeTransformer(BlockTransformerBase, ABC):
         transformer in the list is trivial.
     """
 
-    def __init__(self, block_transformer_list, rescale_mode=None,
-                 drop_features=True):
+    def __init__(self, block_transformer_list, rescale_mode=None, drop_features=True):
+        super().__init__()
         self.block_transformer_list = block_transformer_list
         assert len(self.block_transformer_list) > 0, "Need at least one base" \
                                                      "transformer."
@@ -337,6 +476,37 @@ class CompositeTransformer(BlockTransformerBase, ABC):
                 break
         self.rescale_mode = rescale_mode
         self.drop_features = drop_features
+        self._rescale_factors = {}
+        self._trivial_block_indices = {}
+
+    def _fit_one_feature(self, X, k):
+        data_blocks = []
+        for block_transformer in self.block_transformer_list:
+            data_block = block_transformer.fit_transform_one_feature(
+                X, k, center=False, normalize=False)
+            data_blocks.append(data_block)
+
+        # Handle trivial blocks
+        self._trivial_block_indices[k] = \
+            [idx for idx, data_block in enumerate(data_blocks) if
+             _empty_or_constant(data_block)]
+        if (0 in self._trivial_block_indices[k] and self.drop_features) or \
+                (len(self._trivial_block_indices[k]) == len(data_blocks)):
+            # If first block is trivial and self.drop_features is True,
+            self._centers[k] = np.array([0])
+            self._scales[k] = np.array([1])
+            return
+        else:
+            # Remove trivial blocks
+            for idx in reversed(self._trivial_block_indices[k]):
+                data_blocks.pop(idx)
+        self._rescale_factors[k] = _get_rescale_factors(data_blocks, self.rescale_mode)
+        composite_block = np.hstack(
+            [data_block / scale_factor for data_block, scale_factor in
+             zip(data_blocks, self._rescale_factors[k])]
+        )
+        self._centers[k] = composite_block.mean(axis=0)
+        self._scales[k] = composite_block.std(axis=0)
 
     def _transform_one_feature(self, X, k):
         data_blocks = []
@@ -345,20 +515,49 @@ class CompositeTransformer(BlockTransformerBase, ABC):
                 X, k, center=False, normalize=False)
             data_blocks.append(data_block)
         # Handle trivial blocks
-        trivial_block_indices = \
-            [idx for idx, data_block in enumerate(data_blocks) if
-             _empty_or_constant(data_block)]
-        if (0 in trivial_block_indices and self.drop_features) or \
-                (len(trivial_block_indices) == len(data_blocks)):
+        if (0 in self._trivial_block_indices[k] and self.drop_features) or \
+                (len(self._trivial_block_indices[k]) == len(data_blocks)):
             # If first block is trivial and self.drop_features is True,
             # return empty block
             return np.empty((X.shape[0], 0))
         else:
             # Remove trivial blocks
-            for idx in reversed(trivial_block_indices):
+            for idx in reversed(self._trivial_block_indices[k]):
                 data_blocks.pop(idx)
-        composite_block = np.hstack(_rescale_blocks(data_blocks,
-                                                    self.rescale_mode))
+        composite_block = np.hstack(
+            [data_block / scale_factor for data_block, scale_factor in
+             zip(data_blocks, self._rescale_factors[k])]
+        )
+        return composite_block
+
+    def _fit_transform_one_feature(self, X, k):
+        data_blocks = []
+        for block_transformer in self.block_transformer_list:
+            data_block = block_transformer.fit_transform_one_feature(
+                X, k, center=False, normalize=False)
+            data_blocks.append(data_block)
+        # Handle trivial blocks
+        self._trivial_block_indices[k] = \
+            [idx for idx, data_block in enumerate(data_blocks) if
+             _empty_or_constant(data_block)]
+        if (0 in self._trivial_block_indices[k] and self.drop_features) or \
+                (len(self._trivial_block_indices[k]) == len(data_blocks)):
+            # If first block is trivial and self.drop_features is True,
+            # return empty block
+            self._centers[k] = np.array([0])
+            self._scales[k] = np.array([1])
+            return np.empty((X.shape[0], 0))
+        else:
+            # Remove trivial blocks
+            for idx in reversed(self._trivial_block_indices[k]):
+                data_blocks.pop(idx)
+        self._rescale_factors[k] = _get_rescale_factors(data_blocks, self.rescale_mode)
+        composite_block = np.hstack(
+            [data_block / scale_factor for data_block, scale_factor in
+             zip(data_blocks, self._rescale_factors[k])]
+        )
+        self._centers[k] = composite_block.mean(axis=0)
+        self._scales[k] = composite_block.std(axis=0)
         return composite_block
 
 
@@ -393,22 +592,7 @@ def _update_n_node_samples(tree, X):
         tree.tree_.n_node_samples[i] = new_n_node_samples[i]
 
 
-def _center_and_normalize(data_block, center, normalize):
-    if center:
-        data_block -= data_block.mean(axis=0)
-    if normalize:
-        std = data_block.std(axis=0)
-        if any(std == 0):
-            raise Warning("No recaling done."
-                          "At least one feature is constant.")
-        else:
-            data_block_mean = data_block.mean(axis=0)
-            data_block = (data_block - data_block_mean) / \
-                         data_block.std(axis=0) + data_block_mean
-    return data_block
-
-
-def _rescale_blocks(data_blocks, rescale_mode):
+def _get_rescale_factors(data_blocks, rescale_mode):
     if rescale_mode == "max":
         scale_factors = np.array([max(data_block.std(axis=0)) for
                                   data_block in data_blocks])
@@ -420,11 +604,23 @@ def _rescale_blocks(data_blocks, rescale_mode):
     else:
         raise ValueError("Invalid rescale mode.")
     scale_factors = scale_factors / scale_factors[0]
-    rescaled_data_blocks = \
-        [data_block / scale_factor for data_block, scale_factor in
-         zip(data_blocks, scale_factors)]
-    return rescaled_data_blocks
+    return scale_factors
 
 
 def _empty_or_constant(data_block):
     return data_block.shape[1] == 0 or max(data_block.std(axis=0)) == 0
+
+
+def _blocked_train_test_split(blocked_data, y, oob_seed):
+    n_samples = len(y)
+    train_indices = _generate_sample_indices(oob_seed, n_samples, n_samples)
+    test_indices = _generate_unsampled_indices(oob_seed, n_samples, n_samples)
+    train_blocked_data, test_blocked_data = \
+        blocked_data.train_test_split(train_indices, test_indices)
+    if y.ndim > 1:
+        y_train = y[train_indices, :]
+        y_test = y[test_indices, :]
+    else:
+        y_train = y[train_indices]
+        y_test = y[test_indices]
+    return train_blocked_data, test_blocked_data, y_train, y_test, train_indices, test_indices
