+++ imodels/importance/gmdi.py
@@ -1,46 +1,34 @@
-import copy
-
 import numpy as np
 import pandas as pd
-import warnings
-from sklearn.ensemble._forest import _generate_unsampled_indices, \
-    _generate_sample_indices
-from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
-from sklearn.metrics import log_loss
-from sklearn.preprocessing import OneHotEncoder
+from scipy.spatial.distance import pdist
+from functools import partial
 
-from .block_transformers import GmdiDefaultTransformer, TreeTransformer
-from .ppms import RidgePPM, LogisticPPM
+from .ppms import PartialPredictionModelBase, GenericRegressorPPM, GenericClassifierPPM
+from .block_transformers import _blocked_train_test_split
+from .ranking_stability import tauAP_b, rbo
 
 
-class GMDI:
+class ForestGMDI:
     """
-    The class object for computing GMDI feature importances. Generalized mean
-    decrease in impurity (GMDI) is a flexible framework for computing RF
+    The class object for computing GMDI feature importances for a forest or collection of trees.
+    Generalized mean decrease in impurity (GMDI) is a flexible framework for computing RF
     feature importances. For more details, refer to [paper].
 
     Parameters
     ----------
-    rf_model: scikit-learn random forest object or None
-        The RF model to be used for interpretation. If None, then a new
-        RandomForestRegressor or RandomForestClassifier is instantiated with
-        **kwargs, depending on the task flag.
-    partial_prediction_model: A _PartialPredictionModelBase object or "auto"
-        The partial prediction model to be used for computing partial
-        predictions.
-        If value is set to "auto", then a default is chosen as follows:
-         - If task is set to "regression", then RidgePPM is used.
-         - If task is set to "classification", then LogisticPPM is used.
-    scoring_fns: function or dict with functions as values or "auto"
+    estimators: list of fitted PartialPredictionModelBase objects or scikit-learn type estimators
+        The fitted partial prediction models (one per tree) to use for evaluating
+        feature importance via GMDI. If not a PartialPredictionModelBase, then
+        the estimator is coerced into a PartialPredictionModelBase object via
+        GenericRegressorPPM or GenericClassifierPPM depending on the specified
+        task. Note that these generic PPMs may be computationally expensive.
+    transformers: list of BlockTransformerBase objects
+        The block feature transformers used to generate blocks of engineered
+        features for each original feature. The transformed data is then used
+        as input into the partial prediction models. Should be the same length
+        as estimators.
+    scoring_fns: a function or dict with functions as value and function name (str) as key
         The scoring functions used for evaluating the partial predictions.
-        If "auto", then a default is chosen as follows:
-         - If task is set to "regression", then r2_score is used.
-         - If task is set to "classification", then log_loss is used.
-    mode: string in {"keep_k", "keep_rest"}
-        Mode for the method. "keep_k" imputes the mean of each feature not
-        in block k when making a partial model prediction, while "keep_rest"
-        imputes the mean of each feature in block k. "keep_k" is strongly
-        recommended for computational considerations.
     sample_split: string in {"loo", "oob", "inbag"} or None
         The sample splitting strategy to be used when evaluating the partial
         model predictions. The default "loo" (leave-one-out) is strongly
@@ -49,58 +37,51 @@ class GMDI:
         also be used to overcome these biases. "inbag" is the sample splitting
         strategy used by MDI. If None, no sample splitting is performed and the
         full data set is used to evaluate the partial model predictions.
-    include_raw: bool
-        Flag for whether to augment the local decision stump features extracted
-        from the RF model with the original features.
-    drop_features: bool
-        Flag for whether to use an intercept model for the partial predictions
-        on a given feature if a tree does not have any nodes that split on it,
-        or to use a model on the raw feature (if include_raw is True).
-    refit_rf: bool
-        Flag for whether to refit the supplied RF model when fitting the GMDI
-        feature importances.
-    task: string in {"regression", "classifcation"}
+    tree_random_states: list of int or None
+        Random states from each tree in the fitted random forest; used in
+        sample splitting and only required if sample_split = "oob" or "inbag".
+        Should be the same length as estimators.
+    mode: string in {"keep_k", "keep_rest"}
+        Mode for the method. "keep_k" imputes the mean of each feature not
+        in block k when making a partial model prediction, while "keep_rest"
+        imputes the mean of each feature in block k. "keep_k" is strongly
+        recommended for computational considerations.
+    task: string in {"regression", "classification"}
         The supervised learning task for the RF model. Used for choosing
-        defaults for partial_prediction_model and scoring_fns. Currently only
-        regression and classification are supported.
+        defaults for the scoring_fns. Currently only regression and
+        classification are supported.
+    center: bool
+        Flag for whether to center the transformed data in the transformers.
+    normalize: bool
+        Flag for whether to rescale the transformed data to have unit
+        variance in the transformers.
     """
 
-    def __init__(self,
-                 rf_model=None,
-                 partial_prediction_model="auto",
-                 scoring_fns="auto",
-                 mode="keep_k",
-                 sample_split="loo",
-                 include_raw=True,
-                 drop_features=True,
-                 refit_rf=True,
-                 task="regression",
-                 **kwargs):
-        assert mode in ["keep_k", "keep_rest"]
+    def __init__(self, estimators, transformers, scoring_fns,
+                 sample_split="loo", tree_random_states=None, mode="keep_k",
+                 task="regression", center=True, normalize=False):
         assert sample_split in ["loo", "oob", "inbag", None]
+        assert mode in ["keep_k", "keep_rest"]
         assert task in ["regression", "classification"]
-        if rf_model is not None:
-            self.rf_model = copy.deepcopy(rf_model)
-        elif task == "regression":
-            self.rf_model = RandomForestRegressor(**kwargs)
-        elif task == "classification":
-            self.rf_model = RandomForestClassifier(**kwargs)
-        else:
-            raise ValueError("Unsupported task.")
-        self.partial_prediction_model = partial_prediction_model
+        self.estimators = estimators
+        self.transformers = transformers
         self.scoring_fns = scoring_fns
-        self.mode = mode
         self.sample_split = sample_split
-        self.include_raw = include_raw
-        self.drop_features = drop_features
-        self.refit_rf = refit_rf
+        self.tree_random_states = tree_random_states
+        if self.sample_split in ["oob", "inbag"] and not self.tree_random_states:
+            raise ValueError("Must specify tree_random_states to use 'oob' or 'inbag' sample_split.")
+        self.mode = mode
         self.task = task
+        self.center = center
+        self.normalize = normalize
         self.is_fitted = False
-        self._scores = pd.DataFrame({})
+        self.prediction_score_ = pd.DataFrame({})
+        self.feature_importances_ = pd.DataFrame({})
+        self.feature_importances_by_tree_ = {}
 
-    def get_scores(self, X=None, y=None):
+    def get_scores(self, X, y):
         """
-        Obtain the GMDI feature importances.
+        Obtain the GMDI feature importances for a forest.
 
         Parameters
         ----------
@@ -115,101 +96,128 @@ class GMDI:
         scores: pd.DataFrame of shape (n_features, n_scoring_fns)
             The GMDI feature importances.
         """
-        if self.is_fitted:
-            pass
+        self._fit_importance_scores(X, y)
+        return self.feature_importances_
+
+    def get_stability_scores(self, B=10, metrics="auto"):
+        """
+        Evaluate the stability of the GMDI feature importance rankings
+        across bootstrapped samples of trees. Can be used to select the GLM
+        and scoring metric in a data-driven manner, where the GLM and metric that
+        yields the most stable feature rankings across bootstrapped samples is selected.
+
+        Parameters
+        ----------
+        B: int
+            Number of bootstrap samples.
+        metrics: "auto" or a dict with functions as value and function name (str) as key
+            Metric(s) used to evaluate the stability between two sets of feature importances.
+            If "auto", then the feature importance stability metrics are:
+                (1) Rank-based overlap (RBO) with p=0.9 (from "A Similarity Measure for
+                Indefinite Rankings" by Webber et al. (2010)). Intuitively, this metric gives
+                more weight to features with the largest importances, with most of the weight
+                going to the ~1/(1-p) features with the largest importances.
+                (2) A weighted kendall tau metric (tauAP_b from "The Treatment of Ties in
+                AP Correlation" by Urbano and Marrero (2017)), which also gives more weight
+                to the features with the largest importances, but uses a different weighting
+                scheme from RBO.
+            Note that these default metrics assume that a higher GMDI score indicates
+            greater importance and thus give more weight to these features with high
+            importance/ranks. If a lower GMDI score indicates higher importance, then invert
+            either these stability metrics or the GMDI scores before evaluating the stability.
+
+        Returns
+        -------
+        stability_results: pd.DataFrame of shape (n_features, n_metrics)
+            The stability scores of the GMDI feature rankings across bootstrapped samples.
+
+        """
+        if metrics == "auto":
+            metrics = {"RBO": partial(rbo, p=0.9), "tauAP": tauAP_b}
+        elif not isinstance(metrics, dict):
+            raise ValueError("`metrics` must be 'auto' or a dictionary "
+                             "where the key is the metric name and the value is the evaluation function")
+        single_scoring_fn = not isinstance(self.feature_importances_by_tree_, dict)
+        if single_scoring_fn:
+            feature_importances_dict = {"gmdi_score": self.feature_importances_by_tree_}
         else:
-            if X is None or y is None:
-                raise ValueError("Not yet fitted. Need X and y as inputs.")
-            else:
-                self._fit_importance_scores(X, y)
-        return self._scores
+            feature_importances_dict = self.feature_importances_by_tree_
+        stability_dict = {}
+        for scoring_fn_name, feature_importances_by_tree in feature_importances_dict.items():
+            n_trees = feature_importances_by_tree.shape[1]
+            fi_scores_boot_ls = []
+            for b in range(B):
+                bootstrap_sample = np.random.choice(n_trees, n_trees, replace=True)
+                fi_scores_boot_ls.append(feature_importances_by_tree[bootstrap_sample].mean(axis=1))
+            fi_scores_boot = pd.concat(fi_scores_boot_ls, axis=1)
+            stability_results = {"scorer": [scoring_fn_name]}
+            for metric_name, metric_fun in metrics.items():
+                stability_results[metric_name] = [np.mean(pdist(fi_scores_boot.T, metric=metric_fun))]
+            stability_dict[scoring_fn_name] = pd.DataFrame(stability_results)
+        stability_df = pd.concat(stability_dict, axis=0).reset_index(drop=True)
+        if single_scoring_fn:
+            stability_df = stability_df.drop(columns=["scorer"])
+        return stability_df
 
     def _fit_importance_scores(self, X, y):
-        if self.refit_rf:
-            self.rf_model.fit(X, y)
         all_scores = []
-        if self.task == "regression":
-            ppm, scoring_fns = self._get_reg_defaults()
-        elif self.task == "classification":
-            ppm, scoring_fns = self._get_classification_defaults()
-            if len(np.unique(y)) > 2:
-                y = OneHotEncoder().fit_transform(y.reshape(-1, 1)).toarray()
-        else:
-            raise ValueError("Only regression and classification tasks "
-                             "currently supported.")
-            # ppm, scoring_fns = self.partial_prediction_model, self.scoring_fns
-        for tree_model in self.rf_model.estimators_:
-            transformer = GmdiDefaultTransformer(
-                tree_model, drop_features=self.drop_features) if \
-                self.include_raw else TreeTransformer(tree_model)
-            if self.sample_split == "loo":
-                ppm.loo = True
-            else:
-                ppm.loo = False
-            gmdi_helper = GmdiHelper(transformer, ppm, scoring_fns,
-                                     self.mode, self.sample_split)
-            scores = gmdi_helper.get_scores(X, y)
+        all_full_preds = []
+        for estimator, transformer, tree_random_state in \
+                zip(self.estimators, self.transformers, self.tree_random_states):
+            tree_gmdi = TreeGMDI(estimator=estimator,
+                                 transformer=transformer,
+                                 scoring_fns=self.scoring_fns,
+                                 sample_split=self.sample_split,
+                                 tree_random_state=tree_random_state,
+                                 mode=self.mode,
+                                 task=self.task,
+                                 center=self.center,
+                                 normalize=self.normalize)
+            scores = tree_gmdi.get_scores(X, y)
             if scores is not None:
                 all_scores.append(scores)
-        scoring_fns = scoring_fns if isinstance(scoring_fns, dict) \
-            else {"importance": scoring_fns}
-        for fn_name in scoring_fns.keys():
-            if len(all_scores) == 0:
-                raise ValueError("Transformer representation was empty for all trees.")
-            self._scores[fn_name] = np.mean([scores[fn_name] for scores
-                                             in all_scores], axis=0)
+                all_full_preds.append(tree_gmdi._full_preds)
+        if len(all_scores) == 0:
+            raise ValueError("Transformer representation was empty for all trees.")
+        full_preds = np.nanmean(all_full_preds, axis=0)
+        self._full_preds = full_preds
+        scoring_fns = self.scoring_fns if isinstance(self.scoring_fns, dict) \
+            else {"importance": self.scoring_fns}
+        for fn_name, scoring_fn in scoring_fns.items():
+            self.feature_importances_by_tree_[fn_name] = pd.concat([scores[fn_name] for scores in all_scores], axis=1)
+            self.feature_importances_by_tree_[fn_name].columns = np.arange(len(all_scores))
+            self.feature_importances_[fn_name] = np.mean(self.feature_importances_by_tree_[fn_name], axis=1)
+            self.prediction_score_[fn_name] = [scoring_fn(y[~np.isnan(full_preds)], full_preds[~np.isnan(full_preds)])]
+        if list(scoring_fns.keys()) == ["importance"]:
+            self.prediction_score_ = self.prediction_score_["importance"]
+            self.feature_importances_by_tree_ = self.feature_importances_by_tree_["importance"]
         if isinstance(X, pd.DataFrame):
-            self._scores.index = X.columns
-        self._scores.index.name = 'var'
-        self._scores.reset_index(inplace=True)
+            self.feature_importances_.index = X.columns
+        self.feature_importances_.index.name = 'var'
+        self.feature_importances_.reset_index(inplace=True)
         self.is_fitted = True
 
-    def _get_reg_defaults(self):
-        if self.partial_prediction_model == "auto":
-            ppm = RidgePPM()
-        else:
-            ppm = self.partial_prediction_model
-        if self.scoring_fns == "auto":
-            scoring_fns = {"importance": _fast_r2_score}
-        else:
-            scoring_fns = self.scoring_fns
-        return ppm, scoring_fns
 
-    def _get_classification_defaults(self):
-        if self.partial_prediction_model == "auto":
-            ppm = LogisticPPM()
-        else:
-            ppm = self.partial_prediction_model
-        if self.scoring_fns == "auto":
-            scoring_fns = {"importance": _neg_log_loss}
-        else:
-            scoring_fns = self.scoring_fns
-        return ppm, scoring_fns
-
-
-class GmdiHelper:
+class TreeGMDI:
     """
-    A class that is primarily used by GMDI for fitting the GMDI scores for
-    a single tree in the RF. This class can also be used for further extensions
-    of the GMDI framework by supplying block transformers that provide
-    other types of feature engineering.
+    The class object for computing GMDI feature importances for a single tree.
+    Generalized mean decrease in impurity (GMDI) is a flexible framework for computing RF
+    feature importances. For more details, refer to [paper].
 
     Parameters
     ----------
-    transformer: A BlockTransformerBase object
+    estimator: a fitted PartialPredictionModelBase object or scikit-learn type estimator
+        The fitted partial prediction model to use for evaluating
+        feature importance via GMDI. If not a PartialPredictionModelBase, then
+        the estimator is coerced into a PartialPredictionModelBase object via
+        GenericRegressorPPM or GenericClassifierPPM depending on the specified
+        task. Note that these generic PPMs may be computationally expensive.
+    transformer: a BlockTransformerBase object
         A block feature transformer used to generate blocks of engineered
-        features for each original feature. GMDI is computed by evaluating
-        partial models on these blocks.
-    partial_prediction_model: A _PartialPredictionModelBase object
-        The partial prediction model to be used for computing partial
-        predictions.
-    scoring_fns: function or dict with functions as values
+        features for each original feature. The transformed data is then used
+        as input into the partial prediction models.
+    scoring_fns: a function or dict with functions as value and function name (str) as key
         The scoring functions used for evaluating the partial predictions.
-    mode: string in {"keep_k", "keep_rest"}
-        Mode for the method. "keep_k" imputes the mean of each feature not
-        in block k when making a partial model prediction, while "keep_rest"
-        imputes the mean of each feature in block k. "keep_k" is strongly
-        recommended for computational considerations.
     sample_split: string in {"loo", "oob", "inbag"} or None
         The sample splitting strategy to be used when evaluating the partial
         model predictions. The default "loo" (leave-one-out) is strongly
@@ -218,33 +226,51 @@ class GmdiHelper:
         also be used to overcome these biases. "inbag" is the sample splitting
         strategy used by MDI. If None, no sample splitting is performed and the
         full data set is used to evaluate the partial model predictions.
+    tree_random_state: int or None
+        Random state of the fitted tree; used in sample splitting and
+        only required if sample_split = "oob" or "inbag".
+    mode: string in {"keep_k", "keep_rest"}
+        Mode for the method. "keep_k" imputes the mean of each feature not
+        in block k when making a partial model prediction, while "keep_rest"
+        imputes the mean of each feature in block k. "keep_k" is strongly
+        recommended for computational considerations.
+    task: string in {"regression", "classification"}
+        The supervised learning task for the RF model. Used for choosing
+        defaults for the scoring_fns. Currently only regression and
+        classification are supported.
     center: bool
-        Flag for whether to center the engineered features.
+        Flag for whether to center the transformed data in the transformers.
     normalize: bool
-        Flag for whether to rescale the engineered features to have unit
-        variance.
+        Flag for whether to rescale the transformed data to have unit
+        variance in the transformers.
     """
 
-    def __init__(self, transformer, partial_prediction_model, scoring_fns,
-                 mode="keep_k", sample_split="loo", center=True, normalize=False):
+    def __init__(self, estimator, transformer, scoring_fns,
+                 sample_split="loo", tree_random_state=None, mode="keep_k",
+                 task="regression", center=True, normalize=False):
+        assert sample_split in ["loo", "oob", "inbag", "auto", None]
         assert mode in ["keep_k", "keep_rest"]
-        assert sample_split in ["loo", "oob", "inbag", None]
+        assert task in ["regression", "classification"]
+        self.estimator = estimator
         self.transformer = transformer
-        self.partial_prediction_model = copy.deepcopy(partial_prediction_model)
         self.scoring_fns = scoring_fns
-        self.mode = mode
         self.sample_split = sample_split
-        if sample_split in ["oob", "inbag"] and hasattr(partial_prediction_model, "loo") and \
-                partial_prediction_model.loo:
-            raise ValueError("Cannot use LOO together with OOB or in-bag sample splitting.")
+        self.tree_random_state = tree_random_state
+        _validate_sample_split(self.sample_split, self.estimator, isinstance(self.estimator, PartialPredictionModelBase))
+        if self.sample_split in ["oob", "inbag"] and not self.tree_random_state:
+            raise ValueError("Must specify tree_random_state to use 'oob' or 'inbag' sample_split.")
+        self.mode = mode
+        self.task = task
         self.center = center
         self.normalize = normalize
-        self._scores = None
         self.is_fitted = False
+        self._full_preds = None
+        self.prediction_score_ = None
+        self.feature_importances_ = None
 
-    def get_scores(self, X=None, y=None):
+    def get_scores(self, X, y):
         """
-        Obtain the GMDI feature importances.
+        Obtain the GMDI feature importances for a single tree.
 
         Parameters
         ----------
@@ -259,78 +285,53 @@ class GmdiHelper:
         scores: pd.DataFrame of shape (n_features, n_scoring_fns)
             The GMDI feature importances.
         """
-        if X is None or y is None:
-            if self.is_fitted:
-                pass
-            else:
-                raise ValueError("Not yet fitted. Need X and y as inputs.")
-        else:
-            self._fit_importance_scores(X, y)
-        return self._scores
+        self._fit_importance_scores(X, y)
+        return self.feature_importances_
 
     def _fit_importance_scores(self, X, y):
+        n_samples = y.shape[0]
         blocked_data = self.transformer.transform(X, center=self.center,
                                                   normalize=self.normalize)
         self.n_features = blocked_data.n_blocks
-        if self.sample_split == "oob":
-            train_blocked_data, test_blocked_data, y_train, y_test = \
-                self._train_test_split(blocked_data, y)
-        elif self.sample_split == "inbag":
-            train_blocked_data, _, y_train, _ = \
-                self._train_test_split(blocked_data, y)
-            test_blocked_data = train_blocked_data
-            y_test = y_train
-        else:
-            train_blocked_data = test_blocked_data = blocked_data
-            y_train = y_test = y
-        if train_blocked_data.get_all_data().shape[1] == 0:
-            self._scores = None
-            # warnings.warn("Transformer representation is empty.")
-        else:
-            full_preds, partial_preds = self._get_preds(
-                train_blocked_data, y_train, test_blocked_data, y_test)
-            self._score_partial_predictions(full_preds, partial_preds, y_test)
+        train_blocked_data, test_blocked_data, y_train, y_test, test_indices = \
+            _get_sample_split_data(blocked_data, y, self.tree_random_state, self.sample_split)
+        if train_blocked_data.get_all_data().shape[1] != 0:
+            if hasattr(self.estimator, "predict_full") and \
+                    hasattr(self.estimator, "predict_partial"):
+                full_preds = self.estimator.predict_full(test_blocked_data)
+                partial_preds = self.estimator.predict_partial(test_blocked_data, mode=self.mode)
+            else:
+                if self.task == "regression":
+                    ppm = GenericRegressorPPM(self.estimator)
+                elif self.task == "classification":
+                    ppm = GenericClassifierPPM(self.estimator)
+                full_preds = ppm.predict_full(test_blocked_data)
+                partial_preds = ppm.predict_partial(test_blocked_data, mode=self.mode)
+            self._score_full_predictions(y_test, full_preds)
+            self._score_partial_predictions(y_test, full_preds, partial_preds)
+
+            full_preds_n = np.empty(n_samples) if full_preds.ndim == 1 \
+                else np.empty((n_samples, full_preds.shape[1]))
+            full_preds_n[:] = np.nan
+            full_preds_n[test_indices] = full_preds
+            self._full_preds = full_preds_n
         self.is_fitted = True
 
-    def _get_preds_one_target(self, train_blocked_data, y_train,
-                              test_blocked_data, y_test):
-        ppm = copy.deepcopy(self.partial_prediction_model)
-        ppm.fit(train_blocked_data, y_train, test_blocked_data,
-                y_test, self.mode)
-        full_preds = ppm.get_full_predictions()
-        partial_preds = ppm.get_all_partial_predictions()
-        return full_preds, partial_preds
+    def _score_full_predictions(self, y_test, full_preds):
+        scoring_fns = self.scoring_fns if isinstance(self.scoring_fns, dict) \
+            else {"score": self.scoring_fns}
+        all_prediction_scores = pd.DataFrame({})
+        for fn_name, scoring_fn in scoring_fns.items():
+            scores = scoring_fn(y_test, full_preds)
+            all_prediction_scores[fn_name] = [scores]
+        self.prediction_score_ = all_prediction_scores
 
-    def _get_preds(self, train_blocked_data, y_train,
-                   test_blocked_data, y_test):
-        if y_train.ndim == 1:
-            return self._get_preds_one_target(train_blocked_data, y_train,
-                                              test_blocked_data, y_test)
-        else:
-            full_preds_list = []
-            partial_preds_list = []
-            for j in range(y_train.shape[1]):
-                yj_train = y_train[:, j]
-                yj_test = y_test[:, j]
-                full_preds_j, partial_preds_j = self._get_preds_one_target(
-                    train_blocked_data, yj_train, test_blocked_data, yj_test)
-                full_preds_list.append(full_preds_j)
-                partial_preds_list.append(partial_preds_j)
-            full_preds = np.array(full_preds_list).T
-            partial_preds = dict()
-            for k in range(self.n_features):
-                partial_preds[k] = \
-                    np.array([partial_preds_j[k] for partial_preds_j in
-                              partial_preds_list]).T
-            return full_preds, partial_preds
-
-    def _score_partial_predictions(self, full_preds, partial_preds, y_test):
+    def _score_partial_predictions(self, y_test, full_preds, partial_preds):
         scoring_fns = self.scoring_fns if isinstance(self.scoring_fns, dict) \
             else {"importance": self.scoring_fns}
         all_scores = pd.DataFrame({})
         for fn_name, scoring_fn in scoring_fns.items():
-            scores = _partial_preds_to_scores(partial_preds, y_test,
-                                              scoring_fn)
+            scores = _partial_preds_to_scores(partial_preds, y_test, scoring_fn)
             if self.mode == "keep_rest":
                 full_score = scoring_fn(y_test, full_preds)
                 scores = full_score - scores
@@ -342,46 +343,44 @@ class GmdiHelper:
                 raise ValueError("Unexpected dimensions. {}".format(msg))
             scores = scores.ravel()
             all_scores[fn_name] = scores
-        self._scores = all_scores
-
-    def _train_test_split(self, blocked_data, y):
-        n_samples = len(y)
-        train_indices = _generate_sample_indices(
-            self.transformer.oob_seed, n_samples, n_samples)
-        test_indices = _generate_unsampled_indices(
-            self.transformer.oob_seed, n_samples, n_samples)
-        train_blocked_data, test_blocked_data = \
-            blocked_data.train_test_split(train_indices, test_indices)
-        if y.ndim > 1:
-            y_train = y[train_indices, :]
-            y_test = y[test_indices, :]
-        else:
-            y_train = y[train_indices]
-            y_test = y[test_indices]
-        return train_blocked_data, test_blocked_data, y_train, y_test
+        self.feature_importances_ = all_scores
 
 
 def _partial_preds_to_scores(partial_preds, y_test, scoring_fn):
     scores = []
     for k, y_pred in partial_preds.items():
-        if not hasattr(y_pred, "__len__"):  # if constant model
-            y_pred = np.ones_like(y_test) * y_pred
-        elif y_test.shape != y_pred.shape:
-            # if constant model for multitarget prediction
-            y_pred = np.array([y_pred] * y_test.shape[0])
+        if isinstance(y_pred, tuple):  # if constant model
+            y_pred = np.ones_like(y_test) * y_pred[1]
         scores.append(scoring_fn(y_test, y_pred))
     return np.vstack(scores)
 
 
-def _fast_r2_score(y_true, y_pred, multiclass=False):
-    numerator = ((y_true - y_pred) ** 2).sum(axis=0, dtype=np.float64)
-    denominator = ((y_true - np.mean(y_true, axis=0)) ** 2). \
-        sum(axis=0, dtype=np.float64)
-    if multiclass:
-        return np.mean(1 - numerator / denominator)
+def _get_default_sample_split(sample_split, prediction_model, is_ppm):
+    if sample_split == "auto":
+        sample_split = "oob"
+        if is_ppm:
+            if prediction_model.loo:
+                sample_split = "loo"
+    return sample_split
+
+
+def _validate_sample_split(sample_split, prediction_model, is_ppm):
+    if sample_split in ["oob", "inbag"] and is_ppm:
+        if prediction_model.loo:
+            raise ValueError("Cannot use LOO together with OOB or in-bag sample splitting.")
+
+
+def _get_sample_split_data(blocked_data, y, random_state, sample_split):
+    if sample_split == "oob":
+        train_blocked_data, test_blocked_data, y_train, y_test, _, test_indices = \
+            _blocked_train_test_split(blocked_data, y, random_state)
+    elif sample_split == "inbag":
+        train_blocked_data, _, y_train, _, test_indices, _ = \
+            _blocked_train_test_split(blocked_data, y, random_state)
+        test_blocked_data = train_blocked_data
+        y_test = y_train
     else:
-        return 1 - numerator / denominator
-
-
-def _neg_log_loss(y_true, y_pred):
-    return -log_loss(y_true, y_pred)
+        train_blocked_data = test_blocked_data = blocked_data
+        y_train = y_test = y
+        test_indices = np.arange(y.shape[0])
+    return train_blocked_data, test_blocked_data, y_train, y_test, test_indices
