+++ imodels/importance/ppms.py
@@ -6,12 +6,13 @@ from functools import partial
 import numpy as np
 import scipy as sp
 
-from sklearn.linear_model import RidgeCV, Ridge, LogisticRegression, \
-    HuberRegressor
+from sklearn.linear_model import RidgeCV, Ridge, \
+    LogisticRegression, HuberRegressor, Lasso
 from sklearn.metrics import log_loss, mean_squared_error
+from scipy.special import softmax
 
 
-class _PartialPredictionModelBase(ABC):
+class PartialPredictionModelBase(ABC):
     """
     An interface for partial prediction models, objects that make use of a
     block partitioned data object, fits a regression or classification model
@@ -23,88 +24,109 @@ class _PartialPredictionModelBase(ABC):
     ----------
     estimator: scikit estimator object
         The regression or classification model used to obtain predictions.
-
     """
 
     def __init__(self, estimator):
         self.estimator = copy.deepcopy(estimator)
-        self.n_blocks = None
-        self._partial_preds = dict({})
-        self._full_preds = None
         self.is_fitted = False
 
-    def fit(self, train_blocked_data, y_train, test_blocked_data, y_test=None,
-            mode="keep_k"):
+    def fit(self, X, y):
         """
-        Fit the partial predictions model. The regression or classification
-        model is fit on the training data, and then the predictions are made
-        and scored on the test data. Test and training data can be the same.
+        Fit the partial prediction model.
 
         Parameters
         ----------
-        train_blocked_data: BlockPartitionedData object
-            Training covariate data
-        y_train: ndarray of shape (n_samples, n_targets)
-            Training response data
-        test_blocked_data: BlockPartitionedData object
-            Test covariate data
-        y_test: ndarray of shape (n_samples, n_targets)
-            Training response data
+        X: ndarray of shape (n_samples, n_features)
+            The covariate matrix.
+        y: ndarray of shape (n_samples, n_targets)
+            The observed responses.
+        """
+        self._fit_model(X, y)
+        self.is_fitted = True
+
+    @abstractmethod
+    def _fit_model(self, X, y):
+        """
+        Fit the regression or classification model on all the data.
+
+        Parameters
+        ----------
+        X: ndarray of shape (n_samples, n_features)
+            The covariate matrix.
+        y: ndarray of shape (n_samples, n_targets)
+            The observed responses.
+        """
+        pass
+
+    @abstractmethod
+    def predict(self, X):
+        """
+        Make predictions on new data using the fitted model.
+
+        Parameters
+        ----------
+        X: ndarray of shape (n_samples, n_features)
+            The covariate matrix, for which to make predictions.
+        """
+        pass
+
+    @abstractmethod
+    def predict_full(self, blocked_data):
+        """
+        Make predictions using all the data based upon the fitted model.
+        Used to make full predictions in GMDI.
+
+        Parameters
+        ----------
+        blocked_data: BlockPartitionedData object
+            The block partitioned covariate data, for which to make predictions.
+        """
+        pass
+
+    @abstractmethod
+    def predict_partial_k(self, blocked_data, k, mode):
+        """
+        Make predictions on modified copies of the data based on the fitted model,
+        for a particular feature k of interest. Used to get partial predictions
+        for feature k in GMDI.
+
+        Parameters
+        ----------
+        blocked_data: BlockPartitionedData object
+            The block partitioned covariate data, for which to make predictions.
+        k: int
+            Index of feature in X of interest.
+        mode: string in {"keep_k", "keep_rest"}
+            Mode for the method. "keep_k" imputes the mean of each feature not
+            in block k, "keep_rest" imputes the mean of each feature in block k
+        """
+        pass
+
+    def predict_partial(self, blocked_data, mode):
+        """
+        Make predictions on modified copies of the data based on the fitted model,
+        for each feature under study. Used to get partial predictions in GMDI.
+
+        Parameters
+        ----------
+        blocked_data: BlockPartitionedData object
+            The block partitioned covariate data, for which to make predictions.
         mode: string in {"keep_k", "keep_rest"}
             Mode for the method. "keep_k" imputes the mean of each feature not
             in block k, "keep_rest" imputes the mean of each feature in block k
 
+        Returns
+        -------
+        List of length n_features of partial predictions for each feature.
         """
-        self.n_blocks = train_blocked_data.n_blocks
-        self._fit_model(train_blocked_data, y_train)
-        self._full_preds = \
-            self._fit_full_predictions(test_blocked_data, y_test)
-        for k in range(self.n_blocks):
-            self._partial_preds[k] = \
-                self._fit_partial_predictions(k, mode, test_blocked_data,
-                                              y_test)
-        self.is_fitted = True
-
-    @abstractmethod
-    def _fit_model(self, train_blocked_data, y_train):
-        """
-        Fit the regression or classification model on all the data.
-        """
-        pass
-
-    @abstractmethod
-    def _fit_full_predictions(self, test_blocked_data, y_test=None):
-        """
-        Calculate the predictions made by model on all the data.
-        """
-        pass
-
-    @abstractmethod
-    def _fit_partial_predictions(self, k, mode, test_blocked_data,
-                                 y_test=None):
-        """
-        Calculate the predictions made by the model on modified copies of the
-        data.
-        """
-        pass
-
-    def get_full_predictions(self):
-        """
-        Get the predictions made by model on all the data.
-        """
-        return self._full_preds
-
-    def get_partial_predictions(self, k):
-        """
-        Get the predictions made by the model on modified copies of the data.
-        """
-        return self._partial_preds[k]
-
-    def get_all_partial_predictions(self):
-        return self._partial_preds
+        n_blocks = blocked_data.n_blocks
+        partial_preds = {}
+        for k in range(n_blocks):
+            partial_preds[k] = self.predict_partial_k(blocked_data, k, mode)
+        return partial_preds
 
 
-class GenericPPM(_PartialPredictionModelBase, ABC):
+class _GenericPPM(PartialPredictionModelBase, ABC):
     """
     Partial prediction model for arbitrary estimators. May be slow.
     """
@@ -112,28 +134,41 @@ class GenericPPM(_PartialPredictionModelBase, ABC):
     def __init__(self, estimator):
         super().__init__(estimator)
 
-    def _fit_model(self, train_blocked_data, y_train):
-        self.estimator.fit(train_blocked_data.get_all_data(), y_train)
+    def _fit_model(self, X, y):
+        self.estimator.fit(X, y)
 
-    def _fit_full_predictions(self, test_blocked_data, y_test=None):
-        pred_func = self._get_pred_func()
-        return pred_func(test_blocked_data.get_all_data())
+    def predict(self, X):
+        return self.estimator.predict(X)
 
-    def _fit_partial_predictions(self, k, mode, test_blocked_data,
-                                 y_test=None):
-        pred_func = self._get_pred_func()
-        modified_data = test_blocked_data.get_modified_data(k, mode)
-        return pred_func(modified_data)
+    def predict_full(self, blocked_data):
+        return self.predict(blocked_data.get_all_data())
 
-    def _get_pred_func(self):
-        if hasattr(self.estimator, "predict_proba"):
-            pred_func = self.estimator.predict_proba
-        else:
-            pred_func = self.estimator.predict
-        return pred_func
+    def predict_partial_k(self, blocked_data, k, mode):
+        modified_data = blocked_data.get_modified_data(k, mode)
+        return self.predict(modified_data)
 
 
-class GlmPPM(_PartialPredictionModelBase, ABC):
+class GenericRegressorPPM(_GenericPPM, PartialPredictionModelBase, ABC):
+    """
+    Partial prediction model for arbitrary regression estimators. May be slow.
+    """
+    ...
+
+
+class GenericClassifierPPM(_GenericPPM, PartialPredictionModelBase, ABC):
+    """
+    Partial prediction model for arbitrary classification estimators. May be slow.
+    """
+
+    def predict_proba(self, X):
+        return self.estimator.predict_proba(X)
+
+    def predict_partial_k(self, blocked_data, k, mode):
+        modified_data = blocked_data.get_modified_data(k, mode)
+        return self.predict_proba(modified_data)
+
+
+class _GlmPPM(PartialPredictionModelBase, ABC):
     """
     PPM class for GLM estimator. The GLM estimator is assumed to have a single
     regularization hyperparameter accessible as a named attribute called either
@@ -189,75 +224,106 @@ class GlmPPM(_PartialPredictionModelBase, ABC):
         self.trim = trim
         self.gcv_mode = gcv_mode
         self.hyperparameter_scorer = hyperparameter_scorer
-        self.alpha_ = None
-        self.loo_coefficients_ = None
-        self.coefficients_ = None
+        self.alpha_ = {}
+        self.loo_coefficients_ = {}
+        self.coefficients_ = {}
         self._intercept_pred = None
 
-    def _fit_model(self, train_blocked_data, y_train):
-        # Compute regularization hyperparameter using approximate LOOCV
-        X_train = train_blocked_data.get_all_data()
-        if isinstance(self.estimator, Ridge):
-            cv = RidgeCV(alphas=self.alpha_grid, gcv_mode=self.gcv_mode)
-            cv.fit(X_train, y_train)
-            self.alpha_ = cv.alpha_
-        else:
-            self.alpha_ = self._get_aloocv_alpha(X_train, y_train)
-        # Fit the model on the training set and compute the coefficients
-        if self.loo:
-            self.loo_coefficients_ = \
-                self._fit_loo_coefficients(X_train, y_train, self.alpha_)
-        else:
-            self.coefficients_ = \
-                self._fit_coefficients(X_train, y_train, self.alpha_)
+    def _fit_model(self, X, y):
+        y_train = copy.deepcopy(y)
+        if y_train.ndim == 1:
+            y_train = y_train.reshape(-1, 1)
+        self._n_outputs = y_train.shape[1]
+        for j in range(self._n_outputs):
+            yj = y_train[:, j]
+            # Compute regularization hyperparameter using approximate LOOCV
+            if isinstance(self.estimator, Ridge):
+                cv = RidgeCV(alphas=self.alpha_grid, gcv_mode=self.gcv_mode)
+                cv.fit(X, yj)
+                self.alpha_[j] = cv.alpha_
+            else:
+                self.alpha_[j] = self._get_aloocv_alpha(X, yj)
+            # Fit the model on the training set and compute the coefficients
+            if self.loo:
+                self.loo_coefficients_[j] = \
+                    self._fit_loo_coefficients(X, yj, self.alpha_[j])
+                self.coefficients_[j] = _extract_coef_and_intercept(self.estimator)
+            else:
+                self.coefficients_[j] = \
+                    self._fit_coefficients(X, yj, self.alpha_[j])
 
-    def _fit_full_predictions(self, test_blocked_data, y_test=None):
-        X_test = test_blocked_data.get_all_data()
-        if self.loo:
-            preds = _get_preds(X_test, self.loo_coefficients_,
-                               self.inv_link_fn)
+    def predict(self, X):
+        preds_list = []
+        for j in range(self._n_outputs):
+            preds_j = _get_preds(X, self.coefficients_[j], self.inv_link_fn)
+            preds_list.append(preds_j)
+        if self._n_outputs == 1:
+            preds = preds_list[0]
         else:
-            preds = _get_preds(X_test, self.coefficients_,
-                               self.inv_link_fn)
+            preds = np.stack(preds_list, axis=1)
         return _trim_values(preds, self.trim)
 
-    def _fit_partial_predictions(self, k, mode, test_blocked_data,
-                                 y_test=None):
-        if mode == "keep_k":
-            block_indices = test_blocked_data.get_block_indices(k)
-            data_block = test_blocked_data.get_block(k)
-        elif mode == "keep_rest":
-            block_indices = test_blocked_data.get_all_except_block_indices(k)
-            data_block = test_blocked_data.get_all_except_block(k)
+    def predict_loo(self, X):
+        preds_list = []
+        for j in range(self._n_outputs):
+            if self.loo:
+                preds_j = _get_preds(X, self.loo_coefficients_[j], self.inv_link_fn)
+            else:
+                preds_j = _get_preds(X, self.coefficients_[j], self.inv_link_fn)
+            preds_list.append(preds_j)
+        if self._n_outputs == 1:
+            preds = preds_list[0]
         else:
-            raise ValueError("Invalid mode")
-        if len(block_indices) == 0: # If empty block
+            preds = np.stack(preds_list, axis=1)
+        return _trim_values(preds, self.trim)
+
+    def predict_full(self, blocked_data):
+        return self.predict_loo(blocked_data.get_all_data())
+
+    def predict_partial_k(self, blocked_data, k, mode):
+        assert mode in ["keep_k", "keep_rest"]
+        if mode == "keep_k":
+            block_indices = blocked_data.get_block_indices(k)
+            data_block = blocked_data.get_block(k)
+        elif mode == "keep_rest":
+            block_indices = blocked_data.get_all_except_block_indices(k)
+            data_block = blocked_data.get_all_except_block(k)
+        if len(block_indices) == 0:  # If empty block
             return self.intercept_pred
         else:
-            if self.loo:
-                coefs = self.loo_coefficients_[:, block_indices]
-                intercept = self.loo_coefficients_[:, -1]
+            partial_preds_list = []
+            for j in range(self._n_outputs):
+                if self.loo:
+                    coefs = self.loo_coefficients_[j][:, block_indices]
+                    intercept = self.loo_coefficients_[j][:, -1]
+                else:
+                    coefs = self.coefficients_[j][block_indices]
+                    intercept = self.coefficients_[j][-1]
+                partial_preds_j = _get_preds(
+                    data_block, coefs, self.inv_link_fn, intercept
+                )
+                partial_preds_list.append(partial_preds_j)
+            if self._n_outputs == 1:
+                partial_preds = partial_preds_list[0]
             else:
-                coefs = self.coefficients_[block_indices]
-                intercept = self.coefficients_[-1]
-            return _trim_values(_get_preds(data_block, coefs, self.inv_link_fn,
-                                           intercept), self.trim)
+                partial_preds = np.stack(partial_preds_list, axis=1)
+            return _trim_values(partial_preds, self.trim)
 
     @property
     def intercept_pred(self):
         if self._intercept_pred is None:
-            self._intercept_pred = \
-                _trim_values(self.inv_link_fn(self.coefficients_[-1]),
-                             self.trim)
-        return self._intercept_pred
+            self._intercept_pred = np.array([
+                _trim_values(self.inv_link_fn(self.coefficients_[j][-1]), self.trim) \
+                for j in range(self._n_outputs)
+            ])
+        return ("constant_model", self._intercept_pred)
 
     def _fit_coefficients(self, X, y, alpha):
         _set_alpha(self.estimator, alpha)
         self.estimator.fit(X, y)
-        self.coefficients_ = _extract_coef_and_intercept(self.estimator)
-        return self.coefficients_
+        return _extract_coef_and_intercept(self.estimator)
 
-    def _fit_loo_coefficients(self, X, y, alpha):
+    def _fit_loo_coefficients(self, X, y, alpha, max_h=1-1e-4):
         """
         Get the coefficient (and intercept) for each LOO model. Since we fit
         one model for each sample, this gives an ndarray of shape (n_samples,
@@ -266,6 +332,11 @@ class GlmPPM(_PartialPredictionModelBase, ABC):
         orig_coef_ = self._fit_coefficients(X, y, alpha)
         X1 = np.hstack([X, np.ones((X.shape[0], 1))])
         orig_preds = _get_preds(X, orig_coef_, self.inv_link_fn)
+        support_idxs = orig_coef_ != 0
+        if not any(support_idxs):
+            return orig_coef_ * np.ones_like(X1)
+        X1 = X1[:, support_idxs]
+        orig_coef_ = orig_coef_[support_idxs]
         l_doubledot_vals = self.l_doubledot(y, orig_preds)
         J = X1.T * l_doubledot_vals @ X1
         if self.r_doubledot is not None:
@@ -276,8 +347,13 @@ class GlmPPM(_PartialPredictionModelBase, ABC):
             J += alpha * reg_curvature
         normal_eqn_mat = np.linalg.inv(J) @ X1.T
         h_vals = np.sum(X1.T * normal_eqn_mat, axis=0) * l_doubledot_vals
+        h_vals[h_vals == 1] = max_h
         loo_coef_ = orig_coef_[:, np.newaxis] + \
                     normal_eqn_mat * self.l_dot(y, orig_preds) / (1 - h_vals)
+        if not all(support_idxs):
+            loo_coef_dense_ = np.zeros((X.shape[1] + 1, X.shape[0]))
+            loo_coef_dense_[support_idxs, :] = loo_coef_
+            loo_coef_ = loo_coef_dense_
         return loo_coef_.T
 
     def _get_aloocv_alpha(self, X, y):
@@ -288,11 +364,35 @@ class GlmPPM(_PartialPredictionModelBase, ABC):
             sample_scores = np.sum(loo_coef_ * X1, axis=1)
             preds = _trim_values(self.inv_link_fn(sample_scores), self.trim)
             cv_scores[i] = self.hyperparameter_scorer(y, preds)
-        self.alpha_ = self.alpha_grid[np.argmin(cv_scores)]
-        return self.alpha_
+        return self.alpha_grid[np.argmin(cv_scores)]
 
 
-class RidgePPM(GlmPPM, ABC):
+class GlmRegressorPPM(_GlmPPM, PartialPredictionModelBase, ABC):
+    """
+    PPM class for GLM regression estimator.
+    """
+    ...
+
+
+class GlmClassifierPPM(_GlmPPM, PartialPredictionModelBase, ABC):
+    """
+    PPM class for GLM classification estimator.
+    """
+
+    def predict_proba(self, X):
+        probs = self.predict(X)
+        if probs.ndim == 1:
+            probs = np.stack([1 - probs, probs], axis=1)
+        return probs
+
+    def predict_proba_loo(self, X):
+        probs = self.predict_loo(X)
+        if probs.ndim == 1:
+            probs = np.stack([1 - probs, probs], axis=1)
+        return probs
+
+
+class _RidgePPM(_GlmPPM, PartialPredictionModelBase, ABC):
     """
     PPM class that uses ridge as the estimator.
 
@@ -323,9 +423,36 @@ class RidgePPM(GlmPPM, ABC):
         self.alpha_grid = alphas
 
 
-class LogisticPPM(GlmPPM, ABC):
+class RidgeRegressorPPM(_RidgePPM, GlmRegressorPPM,
+                        PartialPredictionModelBase, ABC):
     """
-    PPM class that uses logistic regression as the estimator.
+    PPM class for regression that uses ridge as the GLM estimator.
+    """
+    ...
+
+
+class RidgeClassifierPPM(_RidgePPM, GlmClassifierPPM,
+                         PartialPredictionModelBase, ABC):
+    """
+    PPM class for classification that uses ridge as the GLM estimator.
+    """
+
+    def predict_proba(self, X):
+        probs = softmax(self.predict(X))
+        if probs.ndim == 1:
+            probs = np.stack([1 - probs, probs], axis=1)
+        return probs
+
+    def predict_proba_loo(self, X):
+        probs = softmax(self.predict_loo(X))
+        if probs.ndim == 1:
+            probs = np.stack([1 - probs, probs], axis=1)
+        return probs
+
+
+class LogisticClassifierPPM(GlmClassifierPPM, PartialPredictionModelBase, ABC):
+    """
+    PPM class for classification that uses logistic regression as the estimator.
 
     Parameters
     ----------
@@ -343,18 +470,24 @@ class LogisticPPM(GlmPPM, ABC):
     """
 
     def __init__(self, loo=True, alpha_grid=np.logspace(-2, 3, 25),
-                 max_iter=1000, trim=0.01, **kwargs):
-        super().__init__(LogisticRegression(max_iter=max_iter, **kwargs),
+                 penalty='l2', max_iter=1000, trim=0.01, **kwargs):
+        assert penalty in ['l2', 'l1']
+        if penalty == 'l2':
+            r_doubledot = lambda a: 1
+        elif penalty == 'l1':
+            r_doubledot = None
+        super().__init__(LogisticRegression(penalty=penalty, max_iter=max_iter, **kwargs),
                          loo, alpha_grid,
                          inv_link_fn=sp.special.expit,
                          l_doubledot=lambda a, b: b * (1 - b),
+                         r_doubledot=r_doubledot,
                          hyperparameter_scorer=log_loss,
                          trim=trim)
 
 
-class RobustPPM(GlmPPM, ABC):
+class RobustRegressorPPM(GlmRegressorPPM, PartialPredictionModelBase, ABC):
     """
-    PPM class that uses Huber robust regression as the estimator.
+    PPM class for regression that uses Huber robust regression as the estimator.
 
     Parameters
     ----------
@@ -379,6 +512,24 @@ class RobustPPM(GlmPPM, ABC):
             l_dot=l_dot,
             l_doubledot=l_doubledot,
             hyperparameter_scorer=loss_fn)
+
+
+class LassoRegressorPPM(GlmRegressorPPM, PartialPredictionModelBase, ABC):
+    """
+    PPM class for regression that uses lasso as the estimator.
+
+    Parameters
+    ----------
+    loo: bool
+        Flag for whether to also use LOO calculations for making predictions.
+    alpha_grid: ndarray of shape (n_alphas, )
+        The grid of alpha values for hyperparameter optimization.
+    **kwargs
+        Other Parameters are passed on to Lasso().
+    """
+
+    def __init__(self, loo=True, alpha_grid=np.logspace(-2, 3, 25), **kwargs):
+        super().__init__(Lasso(**kwargs), loo, alpha_grid, r_doubledot=None)
 
 
 def _trim_values(values, trim=None):
@@ -426,6 +577,25 @@ def _get_preds(data_block, coefs, inv_link_fn, intercept=None):
 
 
 def huber_loss(y, preds, epsilon=1.35):
+    """
+    Evaluates Huber loss function.
+
+    Parameters
+    ----------
+    y: array-like of shape (n,)
+        Vector of observed responses.
+    preds: array-like of shape (n,)
+        Vector of estimated/predicted responses.
+    epsilon: float
+        Threshold, determining transition between squared
+        and absolute loss in Huber loss function.
+
+    Returns
+    -------
+    Scalar value, quantifying the Huber loss. Lower loss
+    indicates better fit.
+
+    """
     total_loss = 0
     for i in range(len(y)):
         sample_absolute_error = np.abs(y[i] - preds[i])
